src/requirements.txt
configparser
pandas
requests
sqlalchemy
mysql-connector-python
pymysql
mysqlclient